requirements.txt
asgiref==3.6.0
Django==4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7

djangorestframework~=3.14.0